==> 1_parsing_xml.ipynb <==
"""
Input: XML from Reference Manager.
Output: CSV with dataframe which has columns extracted from xml tags + flag useful.
"""

import pandas as pd
import xml.etree.cElementTree as ET

def get_elem(record, tag):
    element = []
    if record.iterfind(tag):
        for elem in record.iterfind(tag) :
            element.append(elem.text.strip().encode('utf-8'))
    if element:
        return element
    else:
        return [None]

def parse_xml(filename):
    data = list()
    columns = ['filename','author','fulltitle','subtitle','pages',
               'volume','number','keywords','year','abstract']

    tree = ET.ElementTree(file=filename)

    # Count how many elements of type record we can find
    count = 0
    for elem in tree.iter(tag='record'):
        count += 1
    print count

    # Parsing fields into Dataframe
    for record in tree.iter(tag='record'):
        pdf_file = get_elem(record, 'urls/pdf-urls/url')[0]
        authors = get_elem(record, 'contributors/authors/author/style')
        full_title = get_elem(record, 'titles/title/style')[0]
        sub_title = get_elem(record, 'titles/secondary-title/style')[0]
        periodical_fulltitle = get_elem(record, 'periodical/full-title/style')[0]
        pages = get_elem(record, 'pages')[0]
        volume = get_elem(record, 'volume/style')[0]
        number = get_elem(record, 'number/style')[0]
        keywords = get_elem(record, 'keywords/keyword/style')
        year = get_elem(record, 'dates/year')[0]
        location = get_elem(record, 'pub-location/style')[0]
        publisher = get_elem(record, 'publisher/style')[0]
        isbn = get_elem(record, 'isbn/style')[0]
        abstract = get_elem(record, 'abstract/style')[0]

        data.append([pdf_file, authors, full_title, sub_title, pages, 
                     volume, number, keywords, year, abstract])
    
    df = pd.DataFrame(data, columns=columns)   
    
    # Saving Dataframe as csv
    df.to_csv('{0}'.format(filename.replace('.xml','.csv')), index=0)
    
    return df

"""
Here a summary of the data DKG provided:

664 pdfs about kidney (Niere.zip)
An XML with the metadata about the 664 PDFs (niere-zugeordnete-referenzen-xml.xml)
An XML with the metadata about PDFs excluded (niere-ausgeschlossene-xml.xml) 
"""

classifiedfile = '../metadata/niere-example/niere-zugeordnete-referenzen-xml.xml'
excludedfile = '../metadata/niere-example/niere-ausgeschlossene-xml.xml'

classified = parse_xml(classifiedfile)
print classified.shape

excluded = parse_xml(excludedfile)
print excluded.shape

classified.head()

# Add Flag
classified['useful'] = 1
excluded['useful'] = 0

full_df = pd.concat([classified, excluded])

# Add ID for records
full_df['num'] = range(1, len(full_df) + 1)

print full_df.shape

full_df.head()

# save dataframe to be used for classification task
full_df.to_csv('../metadata/niere-example/classification_df.csv', index=0)

# dataframe stats
positive = full_df[full_df.useful == 1]
negative = full_df[full_df.useful == 0]

print """all records: {0}
    positive records: {1} ({2:.1f}% of all records)
    negative records: {3} ({4:.1f}% of all records)""".format(len(full_df), 
                                                              len(positive), float(len(positive))/len(full_df)*100,
                                                              len(negative), float(len(negative))/len(full_df)*100)
