==> 2_features_extraction.ipynb <==
import pandas as pd

# Read dataframe with extracted information
df_niere = pd.read_csv("../metadata/data-dive-files/niere.csv", escapechar='\\')

print df_niere.shape
df_niere.head()

# Select fulltitle and abstract of records
df_niere_valid = df_niere.ix[(df_niere.fulltitle.notnull()) & (df_niere.abstract.notnull()),
                            ["num","fulltitle","abstract","keywords","useful"]]

print df_niere_valid.shape
df_niere_valid.head()

pd.crosstab(df_niere_valid.useful, columns="freq", normalize=True)

# 2) From strings to numerical vectors
# https://radimrehurek.com/gensim/tut1.html#from-strings-to-vectors

import nltk
import gensim

def tokenize(text):
    stopwords = set(nltk.corpus.stopwords.words("english"))
    return [x for x in gensim.utils.tokenize(text, lowercase=True, deacc=True, errors="ignore")
           if x not in stopwords]

df_niere_valid.loc[:,"fulltitle_tokens"] = df_niere_valid["fulltitle"].map(lambda x: tokenize(x))  
df_niere_valid.loc[:,"abstract_tokens"] = df_niere_valid["abstract"].map(lambda x: tokenize(x)) 

df_niere_valid.to_csv("../metadata/niere-example/df_classification_valid.csv", index=0)
df_niere_valid.head()

# get dictionary from fulltitle string
from gensim import corpora

stoplist = set(nltk.corpus.stopwords.words("english"))
documents = list(df_niere_valid["fulltitle"])


# 2a) Tokenize documents: preprocess the strings by removing stopwords, lowering case.

texts = [[word for word in document.lower().split() if word not in stoplist] for document in documents]
           
# 2b) Dictionary token-to-id: only tokens with frequency > 1 into a dictionary
from collections import defaultdict
frequency = defaultdict(int)
for text in texts:
    for token in text:
        frequency[token] += 1

texts = [[token for token in text if frequency[token] > 1]
         for text in texts]

# final dictionary
dictionary = corpora.Dictionary(texts)
dictionary.save('../metadata/niere-example/fulltitle.dict')
print dictionary

# checking tokens
dictionary.token2id

# 2c) Convert tokenized documents from (a) to vector, using dictionary in (b) into a corpus
# tokens not found on dict will be ignored
corpus = [dictionary.doc2bow(text) for text in texts]
corpora.MmCorpus.serialize('../metadata/niere-example/fulltitle.mm', corpus)
