==> 3_feature_transformation.ipynb <==
import pandas as pd
import numpy as np
import nltk
from gensim import corpora, models, similarities, matutils

# load data
df = pd.read_csv("../metadata/niere-example/df_classification_valid.csv")
dictionary = corpora.Dictionary.load("../metadata/niere-example/fulltitle.dict")
corpus = corpora.MmCorpus("../metadata/niere-example/fulltitle.mm")

# 3) Creating a transformation tdf-if https://radimrehurek.com/gensim/tut2.html

# 3a) Initialize a model
# TF-IDF Transformation
tfidf = models.TfidfModel(corpus) 

# 3b) Use the model to transform vectors
corpus_tfidf = tfidf[corpus]

# Available transformations: https://radimrehurek.com/gensim/tut2.html#available-transformations

# 3c) convert a corpus to a standard numpy dense/scipy sparse array
matrix_tfidf = matutils.corpus2dense(corpus=corpus_tfidf,
                                           num_terms=corpus_tfidf.corpus.num_terms,
                                           num_docs=corpus_tfidf.corpus.num_docs)
#corpus 
#num_terms : number of features
#num_docs=None : the corpus length
# matrix with documents as columns
print corpus_tfidf.corpus
print matrix_tfidf.shape

matrix_tfidf_transpose = matrix_tfidf.transpose()
np.savetxt("../metadata/niere-example/tf_idf_matrix.txt", matrix_tfidf_transpose)

print matrix_tfidf_transpose.shape
